Adapt UnitTests to the new namespace and to use interface layer.
No other changes are includes in this commit in order to simplify history.
The System.Web.OData.Test and Microsoft.TestCommon projects now compile.
The WebApiOData.sln now compiles.

All UnitTests are passsing with the exception of TestPublicApi; it requires
additional work in System.Web.OData in order to pass.